+++ services/backend/src/routes/internal/jobs.js
@@ -9,6 +9,7 @@ const { GET_RANDOM_BYTES, hexToBase64 } = require('@lucaapp/crypto');
 
 const database = require('../../database');
 const featureFlag = require('../../utils/featureFlag');
+const { generateNotifications } = require('../../utils/notifications.js');
 
 router.post('/deleteOldTraces', async (request, response) => {
   const t0 = performance.now();
@@ -235,6 +236,15 @@ router.post('/deleteOldTestRedeems', async (request, response) => {
     },
   });
   response.send({ affectedRows, time: performance.now() - t0 });
+});
+
+router.post('/regenerateNotifications', async (request, response) => {
+  const t0 = performance.now();
+  await generateNotifications();
+
+  response.send({
+    time: performance.now() - t0,
+  });
 });
 
 module.exports = router;