+++ services/health-department/src/components/App/modals/HistoryModal/ContactPersonView/Selectors/OverlapTimeSelector.react.js
@@ -36,7 +36,7 @@ export const OverlapTimeSelector = ({ setMinTimeOverlap }) => {
           ) {
             return 0;
           }
-          return value.replace(/[^0-9]/g, '');
+          return value.replace(/\D/g, '');
         }}
         onChange={setMinTimeOverlap}
         autoFocus